+++ DuoArtifacts/docs/release-notes.md
@@ -2,7 +2,8 @@
 
 # Release notes
 
-## Upcoming release
+## 1.5.0
+Released 25 March 2025.
 
 * Changed the Shardblade Infuser (Books-Peri duo) artifact.
 * The Quantum Sanctuary (CAT-Dizzy duo) artifact now leaves you at 1 max shield (up from 0).
@@ -12,6 +13,7 @@
 * Small optimization of artifact rewards.
 
 ## 1.4.0
+Released 16 March 2025.
 
 * Added an option to change how duo artifacts are offered, and changed the default mode: duo artifacts now only get offered once immediately after satisfying their prerequisites, and if not picked, become part of the Common artifact pool for that run. Any other limitations still apply.
 * Clarified how the Frozen Control Rods (Dizzy-Drake duo) artifact works, and changed its implementation to be potentially more compatible with other mods.